+++ examples/gsap/src/components/circles/index.js
@@ -1,2 +1,3 @@
 export { default as ClassCircles } from './class-circles';
-export { default as StatelessCircles } from './stateless-circles';
+export { default as StatelessCirclesHook } from './stateless-circles-hook';
+export { default as StatelessCirclesHoc } from './stateless-circles-hoc';