+++ examples/gsap/src/components/squares/index.js
@@ -1,2 +1,3 @@
 export { default as ClassSquares } from './class-squares';
-export { default as StatelessSquares } from './stateless-squares';
+export { default as StatelessSquaresHook } from './stateless-squares-hook';
+export { default as StatelessSquaresHoc } from './stateless-squares-hoc';