+++ examples/gsap/src/components/usage-variants/stateless-hooks.js
@@ -1,16 +1,16 @@
 import React from 'react';
 
 import Controls from '../controls';
-import { StatelessCirclesHoc } from '../circles';
-import { StatelessSquaresHoc } from '../squares';
+import { StatelessCirclesHook } from '../circles';
+import { StatelessSquaresHook } from '../squares';
 import Halo from '../halo';
 
 const InvalidStateless = () => (
   <div className="app">
     <Halo />
     <Controls />
-    <StatelessCirclesHoc />
-    <StatelessSquaresHoc />
+    <StatelessCirclesHook />
+    <StatelessSquaresHook />
   </div>
 );