+++ src/index.js
@@ -1,2 +1,3 @@
-export { default } from './consumer';
+export { default } from './hook';
 export { RefProvider } from './provider';
+export { default as consumeRefs } from './consumer';