+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - 2019-03-31
+### Changed
+- Hook added, it is default export now
+- HOC kept for use with classes, can be imported as consumeRefs
+- Still work in progress
+
 ## 0.2.2 - 2019-03-18
 ### Changed
 - Babel processing added